requirements.txt
aniso8601==9.0.1
APScheduler==3.9.1
bcc==0.18.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
dnspython==2.2.1
filelock==3.6.0
Flask==2.2.2
flask-ngrok==0.0.25
Flask-RESTful==0.3.9
geographiclib==1.52
geopy==2.2.0
gunicorn
gyp==0.1
hidpidaemon==18.4.6
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
kernelstub==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
MarkupSafe==2.1.1
more-itertools==8.10.0
mysql-connector-python==8.0.30
mysql-connector-python-rf==2.2.2
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
ping3==4.0.3
platformdirs==2.5.1
pop-transition==1.1.2
protobuf==3.12.4
pycairo==1.20.1
pycups==2.0.1
pydbus==0.6.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.2.0
PyNaCl==1.5.0
pyngrok==5.1.0
pyparsing==2.4.7
pypng==0.20220715.0
PyQRCode==1.2.1
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.21.0
python-xlib==0.29
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==5.4.1
repolib==1.5.2
repoman==1.4.0
requests==2.25.1
schedule==1.1.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
speedtest-cli==2.1.3
systemd-python==234
tzdata==2022.2
tzlocal==4.2
ubuntu-advantage-tools==27.10
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==1.26.5
virtualenv==20.13.0+ds
wadllib==1.3.6
Werkzeug==2.2.2
xdg==5
xkit==0.0.0
zipp==1.0.0
